+++ scripts/joiToOpenAPI.js
@@ -70,7 +70,11 @@ async function exec() {
     transformSchemataIn(join(__dirname, '..', 'resource', 'SearchDocumentContentBase2')),
     transformSchemataIn(join(__dirname, '..', 'resource', 'SearchDocument2')),
     transformSchemataIn(join(__dirname, '..', 'money', 'CurrencyCode')),
-    transformSchemataIn(join(__dirname, '..', 'money', 'MonetaryValue'))
+    transformSchemataIn(join(__dirname, '..', 'money', 'MonetaryValue')),
+    transformSchemataIn(join(__dirname, '..', 'money', 'NegativeMonetaryValue')),
+    transformSchemataIn(join(__dirname, '..', 'money', 'NonPositiveMonetaryValue')),
+    transformSchemataIn(join(__dirname, '..', 'money', 'NonNegativeMonetaryValue')),
+    transformSchemataIn(join(__dirname, '..', 'money', 'PositiveMonetaryValue'))
   ])
 }